Split implementation of promise tracer into multiple files.
