+++ CMakeLists.txt
@@ -732,7 +732,7 @@ set(SOURCE_FILES
         src/unix/sys-unix.c
         src/unix/system.c
         src/library/rdt/src/rdt_register_hook.h
-        src/library/rdt/src/r.h)
+        src/library/rdt/src/r.h src/library/rdt/src/rdt_promises/tracer_conf.h src/library/rdt/src/rdt_promises/tracer_state.h src/library/rdt/src/rdt_promises/tracer_output.h src/library/rdt/src/rdt_promises/tracer_sexpinfo.h src/library/rdt/src/rdt_promises/tracer_state.cpp src/library/rdt/src/rdt_promises/tracer_output.cpp src/library/rdt/src/rdt_promises/tracer_sexpinfo.cpp src/library/rdt/src/rdt_promises/tracer_conf.cpp)
 
 include_directories(include)
 include_directories(include/R_ext)
