+++ src/library/rdt/src/Makefile.in
@@ -14,7 +14,10 @@ R_SHARE_DIR = $(R_HOME)/share
 R_INCLUDE_DIR = $(R_HOME)/include
 
 SOURCES_C = init.c rdt.c
-SOURCES_CXX = rdt_promises.cpp rdt_noop.cpp rdt_debug.cpp rdt_trace.cpp
+SOURCES_CXX = rdt_promises.cpp rdt_noop.cpp rdt_debug.cpp rdt_trace.cpp \
+              rdt_promises/tracer_conf.cpp rdt_promises/tracer_output.cpp \
+              rdt_promises/tracer_sexpinfo.cpp rdt_promises/tracer_state.cpp
+
 
 DEPENDS = $(SOURCES_C:.c=.d)
 DEPENDS += $(SOURCES_CXX:.cpp=.d)
